+++ src/single-linked-list.js
@@ -1,4 +1,0 @@
-/**
- * @see {SingleLinkedList}
- */
-module.exports = require('./lib/SingleLinkedList')