+++ src/singly-linked-list.js
@@ -1,0 +1,4 @@
+/**
+ * @see {SinglyLinkedList}
+ */
+module.exports = require('./lib/SinglyLinkedList')